Fixed not being able to select after main menu, updated score to reflect remaining time

+++ Assets/ScoreController.cs
@@ -16,7 +16,7 @@ public class ScoreController : MonoBehaviour {
 
     public void UpdateScore() {
         float score = GameController.Instance.Score;
-        text.text = "" + score;
+        text.text = "" + (int)score;
     }
 
 }
